separating it out into different js files
